Add unit tests for commands and tasks
- Test more functionalities of the program
- Update access level for some functions
- Update build.gradle

+++ src/main/java/seedu/duke/commands/DeleteCommand.java
@@ -8,9 +8,9 @@ import seedu.duke.tasks.TaskManager;
 import java.io.IOException;
 
 public class DeleteCommand extends Command {
-    private int taskNumber;
+    private long taskNumber;
 
-    DeleteCommand(int taskNumber) {
+    DeleteCommand(long taskNumber) {
         this.taskNumber = taskNumber;
     }
